Make sure that the persistent signature is never optimized # missing cast

+++ libhfuzz/main.c
@@ -25,7 +25,7 @@ main(int argc, char** argv) {
      * optimizers
      */
     if (getpid() == -2) {
-      return strlen(LIBHFUZZ_module_main);
+      return (int)strlen(LIBHFUZZ_module_main);
     }
     return HonggfuzzMain(argc, argv);
 }
